refactor(validation): isolate workflow target discovery
